\name{modopt-package}
\alias{modopt-package}
\alias{modopt}
\docType{package}

\title{Modeling Optimization Problems}

\description{
Modeling Optimization Problems
}

\details{
\tabular{ll}{
Package: \tab modopt\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2012-12-21\cr
License: \tab GPL-2\cr
}
}

\author{
Ronald Hochreiter <ron@hochreiter.net>

Maintainer: Ronald Hochreiter <ron@hochreiter.net>
}

\references{ }

\keyword{ package }

\examples{}
